fix: doing a lot of upgrades and updates to the bb method
