+++ tests/test_python_interface.py
@@ -1,6 +1,5 @@
 import inspect
 import os
-import subprocess
 from datetime import datetime, timedelta
 from pathlib import Path
 from unittest import mock
@@ -15,7 +14,6 @@ from copernicusmarine import (
     read_dataframe,
     subset,
 )
-from copernicusmarine.core_functions import exceptions
 
 
 class TestPythonInterface:
@@ -288,135 +286,3 @@ class TestPythonInterface:
         assert "_FillValue" not in subsetdata.depth.attrs
         assert subsetdata.time.attrs["calendar"] == "gregorian"
         assert subsetdata.time.attrs["units"] == "hours since 1950-01-01"
-
-    def test_error_Coord_out_of_dataset_bounds(self):
-        try:
-            _ = subset(
-                dataset_id="cmems_mod_glo_phy_anfc_0.083deg_P1D-m",
-                start_datetime=datetime.today() + timedelta(10),
-                force_download=True,
-                end_datetime=datetime.today()
-                + timedelta(days=10, hours=23, minutes=59),
-            )
-        except exceptions.CoordinatesOutOfDatasetBounds as e:
-            assert "Some of your subset selection" in e.__str__()
-
-    def test_that_requested_interval_fully_included_with_bounding_box_method_outside(
-        self, tmp_path
-    ):
-        output_filename = "output.nc"
-        min_longitude = 0.01
-        max_longitude = 1.55
-        min_latitude = 0.01
-        max_latitude = 1.1
-        min_depth = 30.5
-        max_depth = 50.0
-        start_datetime = "2023-12-01T01:00:00"
-        end_datetime = "2023-12-12T01:00:00"
-        command = [
-            "copernicusmarine",
-            "subset",
-            "--dataset-id",
-            "cmems_mod_glo_phy-thetao_anfc_0.083deg_P1D-m",
-            "--variable",
-            "thetao",
-            "--minimum-longitude",
-            f"{min_longitude}",
-            "--maximum-longitude",
-            f"{max_longitude}",
-            "--minimum-latitude",
-            f"{min_latitude}",
-            "--maximum-latitude",
-            f"{max_latitude}",
-            "--start-datetime",
-            f"{start_datetime}",
-            "--end-datetime",
-            f"{end_datetime}",
-            "--minimum-depth",
-            f"{min_depth}",
-            "--maximum-depth",
-            f"{max_depth}",
-            "--bounding-box-method",
-            "outside",
-            "-o",
-            f"{tmp_path}",
-            "-f",
-            f"{output_filename}",
-            "--force-download",
-        ]
-        output = subprocess.run(command)
-
-        dataset = xarray.open_dataset(Path(tmp_path, output_filename))
-        assert output.returncode == 0
-        assert dataset.longitude.values.min() <= min_longitude
-        assert dataset.longitude.values.max() >= max_longitude
-        assert dataset.latitude.values.min() <= min_latitude
-        assert dataset.latitude.values.max() >= max_latitude
-        assert dataset.depth.values.min() <= min_depth
-        assert dataset.depth.values.max() >= max_depth
-        assert datetime.strptime(
-            str(dataset.time.values.min()), "%Y-%m-%dT%H:%M:%S.000%f"
-        ) <= datetime.strptime(start_datetime, "%Y-%m-%dT%H:%M:%S")
-        assert datetime.strptime(
-            str(dataset.time.values.max()), "%Y-%m-%dT%H:%M:%S.000%f"
-        ) >= datetime.strptime(end_datetime, "%Y-%m-%dT%H:%M:%S")
-
-    def test_that_requested_interval_fully_included_with_bounding_box_method_inside(
-        self, tmp_path
-    ):
-        output_filename = "output.nc"
-        min_longitude = 0.01
-        max_longitude = 1.567
-        min_latitude = 0.013
-        max_latitude = 1.123
-        min_depth = 30.554
-        max_depth = 50.023
-        start_datetime = "2023-12-01T01:00:23"
-        end_datetime = "2023-12-12T01:10:03"
-        command = [
-            "copernicusmarine",
-            "subset",
-            "--dataset-id",
-            "cmems_mod_glo_phy-thetao_anfc_0.083deg_P1D-m",
-            "--variable",
-            "thetao",
-            "--minimum-longitude",
-            f"{min_longitude}",
-            "--maximum-longitude",
-            f"{max_longitude}",
-            "--minimum-latitude",
-            f"{min_latitude}",
-            "--maximum-latitude",
-            f"{max_latitude}",
-            "--start-datetime",
-            f"{start_datetime}",
-            "--end-datetime",
-            f"{end_datetime}",
-            "--minimum-depth",
-            f"{min_depth}",
-            "--maximum-depth",
-            f"{max_depth}",
-            "--bounding-box-method",
-            "inside",
-            "-o",
-            f"{tmp_path}",
-            "-f",
-            f"{output_filename}",
-            "--force-download",
-        ]
-        output = subprocess.run(command)
-
-        dataset = xarray.open_dataset(Path(tmp_path, output_filename))
-        assert output.returncode == 0
-        assert dataset.longitude.values.min() >= min_longitude
-        assert dataset.longitude.values.max() <= max_longitude
-        assert dataset.latitude.values.min() >= min_latitude
-        assert dataset.latitude.values.max() <= max_latitude
-        assert dataset.depth.values.min() >= min_depth
-        assert dataset.depth.values.max() <= max_depth
-        assert datetime.strptime(
-            str(dataset.time.values.min()), "%Y-%m-%dT%H:%M:%S.000%f"
-        ) >= datetime.strptime(start_datetime, "%Y-%m-%dT%H:%M:%S")
-        assert datetime.strptime(
-            str(dataset.time.values.max()), "%Y-%m-%dT%H:%M:%S.000%f"
-        ) <= datetime.strptime(end_datetime, "%Y-%m-%dT%H:%M:%S")
